requirements.txt
aiogram>=3.4.1
google-generativeai>=0.5.2
aiohttp>=3.9.5
pydantic-settings>=2.2.1
